+++ test/2_distribution.test.js
@@ -1,11 +1,8 @@
 const Distribution = artifacts.require('DistributionMock');
 const ERC677BridgeToken = artifacts.require('ERC677BridgeToken');
-const ERC20 = artifacts.require('ERC20');
-const EmptyContract = artifacts.require('EmptyContract');
 
 const { mineBlock } = require('./helpers/ganache');
 
-const ERROR_MSG = 'VM Exception while processing transaction: revert';
 const { BN, toWei } = web3.utils;
 
 require('chai')
@@ -13,73 +10,28 @@ require('chai')
     .use(require('chai-bn')(BN))
     .should();
 
-function calculatePercentage(number, percentage) {
-    return new BN(number).mul(new BN(percentage)).div(new BN(100));
-}
-
 
 contract('Distribution', async accounts => {
-    const TOKEN_NAME = 'DPOS staking token';
-    const TOKEN_SYMBOL = 'DPOS';
-
-    const EMPTY_ADDRESS = '0x0000000000000000000000000000000000000000';
-    const STAKING_EPOCH_DURATION = new BN(604800);
-
-    const REWARD_FOR_STAKING = 1;
-    const ECOSYSTEM_FUND = 2;
-    const PUBLIC_OFFERING = 3;
-    const PRIVATE_OFFERING = 4;
-    const FOUNDATION_REWARD = 5;
-    const EXCHANGE_RELATED_ACTIVITIES = 6;
-
-    const owner = accounts[0];
-
-    const address = {
-        [REWARD_FOR_STAKING]: accounts[1],
-        [ECOSYSTEM_FUND]: accounts[2],
-        [PUBLIC_OFFERING]: accounts[3],
-        [FOUNDATION_REWARD]: accounts[4],
-        [EXCHANGE_RELATED_ACTIVITIES]: accounts[5],
-    };
-
-    const stake = {
-        [REWARD_FOR_STAKING]: new BN(toWei('73000000')),
-        [ECOSYSTEM_FUND]: new BN(toWei('12500000')),
-        [PUBLIC_OFFERING]: new BN(toWei('1000000')),
-        [PRIVATE_OFFERING]: new BN(toWei('8500000')),
-        [FOUNDATION_REWARD]: new BN(toWei('4000000')),
-        [EXCHANGE_RELATED_ACTIVITIES]: new BN(toWei('1000000')),
-    };
-
-    const cliff = {
-        [REWARD_FOR_STAKING]: new BN(12).mul(STAKING_EPOCH_DURATION),
-        [ECOSYSTEM_FUND]: new BN(48).mul(STAKING_EPOCH_DURATION),
-        [PUBLIC_OFFERING]: new BN(0),
-        [PRIVATE_OFFERING]: new BN(4).mul(STAKING_EPOCH_DURATION),
-        [FOUNDATION_REWARD]: new BN(12).mul(STAKING_EPOCH_DURATION),
-    };
-
-    const percentAtCliff = {
-        [ECOSYSTEM_FUND]: 10,
-        [PRIVATE_OFFERING]: 10,
-        [FOUNDATION_REWARD]: 20,
-    };
-
-    const numberOfInstallments = {
-        [ECOSYSTEM_FUND]: new BN(96),
-        [PRIVATE_OFFERING]: new BN(32),
-        [FOUNDATION_REWARD]: new BN(36),
-    };
-
-    const PRIVATE_OFFERING_PRERELEASE = 25; // 25%
-
-    const SUPPLY = new BN(toWei('100000000'));
-
-    let distribution;
-    let token;
-
-    const privateOfferingParticipants = [accounts[6], accounts[7]];
-    const privateOfferingParticipantsStakes = [new BN(toWei('3000000')), new BN(toWei('5500000'))];
+    
+    const {
+        TOKEN_NAME,
+        TOKEN_SYMBOL,
+        STAKING_EPOCH_DURATION,
+        REWARD_FOR_STAKING,
+        ECOSYSTEM_FUND,
+        PUBLIC_OFFERING,
+        PRIVATE_OFFERING,
+        FOUNDATION_REWARD,
+        EXCHANGE_RELATED_ACTIVITIES,
+        address,
+        stake,
+        cliff,
+        percentAtCliff,
+        numberOfInstallments,
+        PRIVATE_OFFERING_PRERELEASE,
+        privateOfferingParticipants,
+        privateOfferingParticipantsStakes,
+    } = require('./constants')(accounts);
 
     function createToken(distributionAddress) {
         return ERC677BridgeToken.new(
@@ -100,6 +52,10 @@ contract('Distribution', async accounts => {
             privateOfferingParticipants,
             privateOfferingParticipantsStakes
         ).should.be.fulfilled;
+    }
+
+    function calculatePercentage(number, percentage) {
+        return new BN(number).mul(new BN(percentage)).div(new BN(100));
     }
 
     function getBalances(addresses) {
